--- a/RecoLocalTracker/SubCollectionProducers/src/PixelClusterSelectorTopBottom.cc
+++ b/RecoLocalTracker/SubCollectionProducers/src/PixelClusterSelectorTopBottom.cc
@@ -4,7 +4,7 @@
 void PixelClusterSelectorTopBottom::produce( edm::StreamID, edm::Event& event, const edm::EventSetup& setup) const {
 
   edm::Handle< SiPixelClusterCollectionNew > input;
-  event.getByToken(label_, input);
+  event.getByToken(token_, input);
 
   edm::ESHandle<TrackerGeometry> geom;
   setup.get<TrackerDigiGeometryRecord>().get( geom );