--- a/RecoLocalTracker/SubCollectionProducers/src/StripClusterSelectorTopBottom.cc
+++ b/RecoLocalTracker/SubCollectionProducers/src/StripClusterSelectorTopBottom.cc
@@ -4,7 +4,7 @@
 void StripClusterSelectorTopBottom::produce( edm::StreamID, edm::Event& event, const edm::EventSetup& setup) const {
 
   edm::Handle< edmNew::DetSetVector<SiStripCluster> > input;
-  event.getByToken(label_, input);
+  event.getByToken(token_, input);
 
   edm::ESHandle<TrackerGeometry> geom;
   setup.get<TrackerDigiGeometryRecord>().get( geom );