--- a/RecoLocalTracker/SubCollectionProducers/src/TrackClusterSplitter.cc
+++ b/RecoLocalTracker/SubCollectionProducers/src/TrackClusterSplitter.cc
@@ -1,5 +1,5 @@
 #include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/Utilities/interface/InputTag.h"
@@ -61,7 +61,7 @@
 #include <memory>
 using namespace std;
 
-class TrackClusterSplitter : public edm::EDProducer 
+class TrackClusterSplitter : public edm::stream::EDProducer<>
 {
 
 public: