--- a/RecoPixelVertexing/PixelLowPtUtilities/plugins/TrackListCombiner.h
+++ b/RecoPixelVertexing/PixelLowPtUtilities/plugins/TrackListCombiner.h
@@ -1,22 +1,31 @@
 #ifndef TrackListCombiner_H
 #define TrackListCombiner_H
 
-#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
+
+#include "TrackingTools/PatternTools/interface/Trajectory.h"
+#include "TrackingTools/PatternTools/interface/TrajTrackAssociation.h"
 
 #include <vector>
 
 namespace edm { class Event; class EventSetup; }
 
-class TrackListCombiner : public edm::EDProducer
+class TrackListCombiner : public edm::global::EDProducer<>
 {
 public:
   explicit TrackListCombiner(const edm::ParameterSet& ps);
   ~TrackListCombiner();
-  virtual void produce(edm::Event& ev, const edm::EventSetup& es);
+  virtual void produce(edm::StreamID, edm::Event& ev, const edm::EventSetup& es) const override;
 
 private:
-  std::vector<std::string> trackProducers;
+  struct Tags {
+    template <typename T1, typename T2> Tags(T1 t1, T2 t2): trajectory(t1), assoMap(t2) {}
+    edm::EDGetTokenT<std::vector<Trajectory>> trajectory;
+    edm::EDGetTokenT<TrajTrackAssociationCollection> assoMap;
+  };
+
+  std::vector<Tags> trackProducers;
 };
 #endif