--- a/RecoPixelVertexing/PixelVertexFinding/src/SealModules.cc
+++ b/RecoPixelVertexing/PixelVertexFinding/src/SealModules.cc
@@ -2,8 +2,6 @@
 #include "FWCore/Framework/interface/MakerMacros.h"
 
 #include "RecoPixelVertexing/PixelVertexFinding/interface/PixelVertexProducer.h"
-#include "RecoPixelVertexing/PixelVertexFinding/interface/SkipBadEvents.h"
 
 
 DEFINE_FWK_MODULE(PixelVertexProducer);
-DEFINE_FWK_MODULE(SkipBadEvents);