--- a/RecoTracker/ConversionSeedGenerators/plugins/PhotonConversionTrajectorySeedProducerFromQuadruplets.cc
+++ b/RecoTracker/ConversionSeedGenerators/plugins/PhotonConversionTrajectorySeedProducerFromQuadruplets.cc
@@ -1,4 +1,4 @@
-#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
 #include "FWCore/Utilities/interface/InputTag.h"
 #include "FWCore/Utilities/interface/Exception.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
@@ -13,7 +13,7 @@
 #include "PhotonConversionTrajectorySeedProducerFromQuadrupletsAlgo.h"
 
 
-class PhotonConversionTrajectorySeedProducerFromQuadruplets : public edm::EDProducer {
+class PhotonConversionTrajectorySeedProducerFromQuadruplets : public edm::stream::EDProducer<> {
 public:
   PhotonConversionTrajectorySeedProducerFromQuadruplets(const edm::ParameterSet& );
   ~PhotonConversionTrajectorySeedProducerFromQuadruplets(){}
@@ -21,7 +21,7 @@ public:
 
 private:
   std::string _newSeedCandidates;
-  PhotonConversionTrajectorySeedProducerFromQuadrupletsAlgo *_theFinder;
+  std::unique_ptr<PhotonConversionTrajectorySeedProducerFromQuadrupletsAlgo> _theFinder;
 };
 
 
@@ -29,7 +29,7 @@ PhotonConversionTrajectorySeedProducerFromQuadruplets::
 PhotonConversionTrajectorySeedProducerFromQuadruplets(const edm::ParameterSet& conf)
   : _newSeedCandidates(conf.getParameter<std::string>( "newSeedCandidates"))
 {
-  _theFinder = new PhotonConversionTrajectorySeedProducerFromQuadrupletsAlgo(conf,
+  _theFinder = std::make_unique<PhotonConversionTrajectorySeedProducerFromQuadrupletsAlgo>(conf,
 	consumesCollector());
   produces<TrajectorySeedCollection>(_newSeedCandidates);