--- a/RecoTracker/NuclearSeedGenerator/plugins/NuclearSeedsEDProducer.cc
+++ b/RecoTracker/NuclearSeedGenerator/plugins/NuclearSeedsEDProducer.cc
@@ -17,7 +17,8 @@ using namespace reco;
 //
 NuclearSeedsEDProducer::NuclearSeedsEDProducer(const edm::ParameterSet& iConfig) : conf_(iConfig),
 improveSeeds(iConfig.getParameter<bool>("improveSeeds")),
-producer_(iConfig.getParameter<std::string>("producer"))
+producer_(consumes<TrajectoryCollection>(iConfig.getParameter<std::string>("producer"))),
+mteToken_(consumes<MeasurementTrackerEvent>(edm::InputTag("MeasurementTrackerEvents")))
 {
    produces<TrajectorySeedCollection>();
    produces<TrajectoryToSeedsMap>();
@@ -42,7 +43,7 @@ NuclearSeedsEDProducer::produce(edm::Event& iEvent, const edm::EventSetup& iSetu
    typedef TrajectoryMeasurement TM;
 
    edm::Handle< TrajectoryCollection > m_TrajectoryCollection;
-   iEvent.getByLabel( producer_, m_TrajectoryCollection );
+   iEvent.getByToken( producer_, m_TrajectoryCollection );
 
    LogDebug("NuclearSeedGenerator") << "Number of trajectory in event :" << m_TrajectoryCollection->size() << "\n";
 
@@ -51,7 +52,7 @@ NuclearSeedsEDProducer::produce(edm::Event& iEvent, const edm::EventSetup& iSetu
 
    
    edm::Handle<MeasurementTrackerEvent> data;
-   iEvent.getByLabel(edm::InputTag("MeasurementTrackerEvent"), data);
+   iEvent.getByToken(mteToken_, data);
 
 //   NavigationSetter setter( *(theNuclearInteractionFinder->nav()) );   why???
 
@@ -90,10 +91,6 @@ NuclearSeedsEDProducer::produce(edm::Event& iEvent, const edm::EventSetup& iSetu
 void
 NuclearSeedsEDProducer::beginRun(edm::Run const& run, const edm::EventSetup& es)
 {
-   theNuclearInteractionFinder = std::auto_ptr<NuclearInteractionFinder>(new NuclearInteractionFinder(es, conf_));
+   theNuclearInteractionFinder = std::make_unique<NuclearInteractionFinder>(es, conf_);
 
 }
-
-void  NuclearSeedsEDProducer::endJob() {}
-
-