… RecoPixelVertexing/PixelVertexFinding/interface/SkipBadEvents.h
… RecoPixelVertexing/PixelVertexFinding/src/SkipBadEvents.cc
… RecoTracker/SpecialSeedGenerators/interface/SimpleCosmicBONSeeder.h
… RecoVertex/AdaptiveVertexFinder/plugins/DoubleVertexFilter.cc
… RecoVertex/NuclearInteractionProducer/interface/NuclearInteractionEDProducer.h
… RecoVertex/NuclearInteractionProducer/plugins/NuclearInteractionEDProducer.cc
… SimGeneral/TrackingAnalysis/interface/SimHitTPAssociationProducer.h
… SimGeneral/TrackingAnalysis/plugins/SimHitTPAssociationProducer.cc
… SimTracker/TrackAssociation/plugins/MCTrackMatcher.cc
… SimTracker/TrackHistory/plugins/GenTrackMatcher.cc
… SimTracker/TrackHistory/plugins/SecondaryVertexTagInfoProxy.cc
… SimTracker/TrackHistory/plugins/TrackCategoriesAnalyzer.cc
… SimTracker/TrackHistory/plugins/TrackingParticleCategoriesAnalyser.cc
… TrackingTools/TrackRefitter/plugins/SealModule.cc
… TrackingTools/TrackRefitter/plugins/TracksToTrajectories.cc
… TrackingTools/TrackRefitter/plugins/TracksToTrajectories.h
… Validation/RecoTrack/plugins/TrackFromSeedProducer.cc
… Validation/RecoVertex/src/MCVerticesAnalyzer.cc
… Validation/RecoVertex/src/MCVerticesWeight.cc
… Validation/RecoVertex/src/MCvsRecoVerticesAnalyzer.cc