Performed also some additional minor cleanup on the same go (unique_ptr, override, consumes, token, etc).

I decided to remove `SkipBadEvents` EDFilter altogether as the comments inside it suggested that it is "obsolete as soon as PoolSource supports skipping events based on event+run numbers", and there were no configuration files using it.

Tested in 7_6_0_pre3, no changes expected.

@rovere @VinInn