@cmsbuild 2015-08-12
A new Pull Request was created by @makortel (Matti Kortelainen) for CMSSW_7_6_X.

Migrate bunch of tracking related modules to one/stream/global

It involves the following packages:

RecoLocalTracker/SubCollectionProducers
RecoPixelVertexing/PixelLowPtUtilities
RecoPixelVertexing/PixelVertexFinding
RecoTracker/ConversionSeedGenerators
RecoTracker/FinalTrackSelectors
RecoTracker/NuclearSeedGenerator
RecoTracker/SpecialSeedGenerators
RecoVertex/AdaptiveVertexFinder
RecoVertex/NuclearInteractionProducer
SimGeneral/TrackingAnalysis
SimTracker/TrackAssociation
SimTracker/TrackHistory
TrackingTools/TrackRefitter
Validation/RecoTrack
Validation/RecoVertex


@civanch, @cvuosalo, @mdhildreth, @cmsbuild, @deguio, @slava77, @danduggan can you please review it and eventually sign? Thanks.
@ghellwig, @forthommel, @yduhm, @nickmccoll, @threus, @venturia, @battibass, @acaudron, @jhgoh, @jlagram, @ferencek, @cerati, @trocino, @abbiendi, @GiacomoSguazzoni, @rovere, @VinInn, @bellan, @wmtford, @mschrode, @dgulhan, @imarches, @gbenelli, @gpetruc, @istaslis, @pvmulder this is something you requested to watch as well.
You can sign-off by replying to this message having '+1' in the first line of your reply.
You can reject by replying  to this message having '-1' in the first line of your reply.
If you are a L2 or a release manager you can ask for tests by saying 'please test' in the first line of a comment.
@degano you are the release manager for this.
You can merge this pull request by typing 'merge' in the first line of your comment.