@slava77 2015-08-12
@cmsbuild please test