@cmsbuild 2015-08-12
-1 
Tested at: 4fed0781949d2dfd04e287f38a48aa8d7c0098b6
 I found an error when building: 
 
 <pre>>> Compiling  /tmp/cmsbuild/workspace/ib-any-integration/CMSSW_7_6_X_2015-08-12-1100/src/TrackingTools/TrackRefitter/src/TrackTransformer.cc 
>> Compiling  /tmp/cmsbuild/workspace/ib-any-integration/CMSSW_7_6_X_2015-08-12-1100/src/TrackingTools/TrackRefitter/src/TrackTransformerForCosmicMuons.cc 
>> Compiling  /tmp/cmsbuild/workspace/ib-any-integration/CMSSW_7_6_X_2015-08-12-1100/src/TrackingTools/TrackRefitter/src/TrackTransformerForGlobalCosmicMuons.cc 
>> Compiling edm plugin /tmp/cmsbuild/workspace/ib-any-integration/CMSSW_7_6_X_2015-08-12-1100/src/TrackingTools/TrackRefitter/plugins/SealModule.cc 
In file included from /tmp/cmsbuild/workspace/ib-any-integration/CMSSW_7_6_X_2015-08-12-1100/src/TrackingTools/TrackRefitter/plugins/SealModule.cc:5:0:
/tmp/cmsbuild/workspace/ib-any-integration/CMSSW_7_6_X_2015-08-12-1100/poison/TrackingTools/TrackRefitter/plugins/TracksToTrajectories.h:1:2: error: #error THIS FILE HAS BEEN REMOVED FROM THE PACKAGE.
 #error THIS FILE HAS BEEN REMOVED FROM THE PACKAGE.
  ^
>> Compiling edm plugin /tmp/cmsbuild/workspace/ib-any-integration/CMSSW_7_6_X_2015-08-12-1100/src/TrackingTools/TrackRefitter/plugins/TracksToTrajectories.cc 
>> Compiling  /tmp/cmsbuild/workspace/ib-any-integration/CMSSW_7_6_X_2015-08-12-1100/src/Validation/RecoTrack/src/MTVHistoProducerAlgo.cc 
>> Compiling  /tmp/cmsbuild/workspace/ib-any-integration/CMSSW_7_6_X_2015-08-12-1100/src/Validation/RecoTrack/src/MTVHistoProducerAlgoFactory.cc 
</pre> 
 you can see the results of the tests here: 
 https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10708/7170/summary.html