@cmsbuild 2015-08-12
Pull request #10708 was updated. @civanch, @cvuosalo, @mdhildreth, @cmsbuild, @deguio, @slava77, @danduggan can you please check and sign again.