@cmsbuild 2015-08-12
The tests are being triggered in jenkins.