@Dr15Jones 2015-08-12
Looks good to me with respect to thread safety.