@makortel 2015-08-12
Thanks @Dr15Jones. I was working to rename the "label" variables with "token", and noticed a compilation warning from TracksToTrajectories.cc:
```
In file included from FWCore/Framework/interface/stream/EDProducerAdaptor.h:25:0,
                 from Framework/interface/stream/EDProducerBase.h:27,
                 from FWCore/Framework/interface/stream/EDProducer.h:28,
                 from TrackingTools/TrackRefitter/plugins/TracksToTrajectories.cc:5:
FWCore/Framework/interface/stream/ProducingModuleAdaptor.h: In instantiation of 'class edm::stream::ProducingModuleAdaptor<TracksToTrajectories, edm::stream::EDProducerBase, edm::stream::EDProducerAdaptorBase>':
FWCore/Framework/interface/stream/EDProducerAdaptor.h:48:83:   required from 'static std::unique_ptr<edm::stream::EDProducerAdaptorBase> edm::MakeModuleHelper<edm::stream::EDProducerAdaptorBase>::makeModule(const edm::ParameterSet&) [with ModType = TracksToTrajectories]'
FWCore/Framework/src/WorkerMaker.h:85:149:   required from 'std::shared_ptr<edm::maker::ModuleHolder> edm::WorkerMaker<T>::makeModule(const edm::ParameterSet&) const [with T = TracksToTrajectories]'
FWCore/Framework/interface/one/outputmoduleAbilityToImplementor.h:66:13:   required from here
FWCore/Framework/interface/stream/ProducingModuleAdaptor.h:37:11: warning: 'edm::stream::ProducingModuleAdaptor<TracksToTrajectories, edm::stream::EDProducerBase, edm::stream::EDProducerAdaptorBase>' has a field 'edm::stream::ProducingModuleAdaptor<TracksToTrajectories, edm::stream::EDProducerBase, edm::stream::EDProducerAdaptorBase>::m_global' whose type uses the anonymous namespace
     class ProducingModuleAdaptor : public B
           ^
```
This seems to be caused by the `GlobalCache`of `TracksToTrajectories` being declared in anonymous namespace (https://github.com/makortel/cmssw/commit/c8abeedc4768ecaf665a28a8f861dcc533184bd2#diff-13c8bed78efca4c92f2dc81a0c379639R27), at least if I move the class back to the global scope, the warning disappears.

Should the GlobalCache class then be declared somewhere else than in anonymous namespace? (if yes, I'll just rename the class to something more specific)