@Dr15Jones 2015-08-12
The warning is because the compiler can't tell if that instantiated templated class is supposed to be seen in more than one .cc file. If it were, you'd actually have two different classes. This is not the case here, but it is impossible for the compiler to tell. That is why it is a warning not an error. I suppose it would be best to be warning free. Rather than rename the class, I'd suggest putting it into a uniquely named namespace.