@makortel 2015-08-12
Hmm, ok. Although then the "problem" of inventing unique class name becomes inventing unique namespace name :) (but I can save some typing with `using namespace ...`)

Do I guess correctly that if the instantiated class template would somehow end up in the anonymous namespace, everything would be fine again? ("somehow" probably being something from difficult to impossible)