@Dr15Jones 2015-08-12
I don't know if everything was in an anonymous namespace if it would stop the complaints.