@cmsbuild 2015-08-12
+1
Tested at: ad3482e49ed8bc6aaad1af4286558fd075ad9d0b
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10708/7190/summary.html