@makortel 2015-08-12
Renamed the "label" variables to "token", and moved the GlobalCache class of TracksToTrajectories to named namespace.