@cmsbuild 2015-08-12
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10708/7190/summary.html