@cmsbuild 2015-08-13
The tests are being triggered in jenkins.