@cvuosalo 2015-08-21
For the record, here are the RECO timing measurements for baseline CMSSW_7_6_0_pre3 with four threads vs. this PR with four threads (no DQM or Validation for either), using 70 events of workflow 25202.0. There appears to be a slight reduction in time, though it is probably not significant.

```
Baseline
Time av 17.0114 s/evt   max 33.375 s on evt 28
M1 Time av 16.8701 s/evt   max 33.375 s on evt 28

PR 10708
Time av 16.9427 s/evt   max 34.0015 s on evt 30
M1 Time av 16.7587 s/evt   max 34.0015 s on evt 30
```